# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Family Budget API for Chege & Lydiah
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: month
          in: query
          schema:
            type: number
        - name: year
          in: query
          schema:
            type: number
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/apply-recurring:
    post:
      operationId: applyRecurringExpenses
      tags: [expenses]
      summary: Copy recurring expenses from the previous month into the current month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyRecurringInput"
      responses:
        "200":
          description: Number of expenses copied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyRecurringResult"
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Monthly budget summary — total budget, total spent, remaining, per-category spent
      parameters:
        - name: month
          in: query
          schema:
            type: number
        - name: year
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
...
  /dashboard/category-breakdown:
    get:
      operationId: getDashboardCategoryBreakdown
      tags: [dashboard]
      summary: Spending breakdown per category for a given month
      parameters:
        - name: month
          in: query
          schema:
            type: number
        - name: year
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"
  /dashboard/trends:
    get:
      operationId: getDashboardTrends
      tags: [dashboard]
      summary: Month-over-month spending totals for the last N months
      parameters:
        - name: months
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Monthly trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthTrend"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Expense:
      type: object
      properties:
        id:
          type: number
        amount:
          type: number
          description: Amount in KES
        category:
          type: string
        description:
          type: string
        notes:
          type: ["string", "null"]
          description: Optional extra notes
        paidById:
          type: string
        paidByName:
          type: string
        isRecurring:
          type: boolean
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - amount
        - category
        - description
        - paidById
        - paidByName
        - isRecurring
        - date
        - createdAt
    ExpenseInput:
      type: object
      properties:
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        notes:
          type: string
        paidById:
          type: string
          description: User ID of who paid — must be a recognised household member
        isRecurring:
          type: boolean
        date:
          type: string
          format: date
      required:
        - amount
        - category
        - description
        - paidById
        - date
    ApplyRecurringInput:
      type: object
      properties:
        month:
          type: number
          description: Target month (1-12)
        year:
          type: number
          description: Target year
      required:
        - month
        - year
    ApplyRecurringResult:
      type: object
      properties:
        copied:
          type: number
          description: Number of expenses copied
      required:
        - copied
...
    DashboardSummary:
      type: object
      properties:
        month:
          type: number
        year:
          type: number
        totalBudget:
          type: number
        totalSpent:
          type: number
        remaining:
          type: number
        chegeContributed:
          type: number
        lydiahContributed:
          type: number
        chegeSpent:
          type: number
        lydiahSpent:
          type: number
        chegeNet:
          type: number
        lydiahNet:
          type: number
        chegeTarget:
          type: number
        lydiahTarget:
          type: number
        expenseCount:
          type: number
      required:
        - month
        - year
        - totalBudget
        - totalSpent
        - remaining
        - chegeContributed
        - lydiahContributed
        - chegeSpent
        - lydiahSpent
        - chegeNet
        - lydiahNet
        - chegeTarget
        - lydiahTarget
        - expenseCount
...
    CategoryBreakdown:
      type: object
      properties:
        category:
          type: string
        budgetAmount:
          type: number
        spentAmount:
          type: number
        remaining:
          type: number
        percentUsed:
          type: number
        priority:
          type: number
        color:
          type: string
      required:
        - category
        - budgetAmount
        - spentAmount
        - remaining
        - percentUsed
        - priority
        - color
    MonthTrend:
      type: object
      properties:
        month:
          type: number
        year:
          type: number
        label:
          type: string
          description: Human-readable label e.g. "Aug 2026"
        totalSpent:
          type: number
        expenseCount:
          type: number
      required:
        - month
        - year
        - label
        - totalSpent
        - expenseCount
...