# lib/api-zod/src/generated/types/applyRecurringInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Family Budget API for Chege & Lydiah
 * OpenAPI spec version: 0.1.0
 */

export interface ApplyRecurringInput {
  /** Target month (1-12) */
  month: number;
  /** Target year */
  year: number;
}